protocol/dataunit: add wire tests
